+++ src/inc/hitable.h
@@ -5,14 +5,14 @@
 #include "ray.h"
 
 typedef struct {
-    float t;
+    double t;
     Vec3  p;
     Vec3  normal;
 } hitrec;
 
 /* puntero a función hit_cb que toma parámetros ... y retorna bool.
  * Declarado como tipo para poder usarlo dentro de un struct */
-typedef bool (*hit_cb)(Ray r, float t_min, float t_max, 
+typedef bool (*hit_cb)(Ray r, double t_min, double t_max, 
                        hitrec *rec, void *self);
 
 typedef struct {
@@ -23,7 +23,7 @@ typedef struct {
 
 void hitable_create(hit_cb, void *);
 
-bool hitable_call_all(Ray, float, float, hitrec *);
+bool hitable_call_all(Ray, double, double, hitrec *);
 
 void hitable_destroy();
