+++ src/sphere.c
@@ -5,7 +5,7 @@
 #include "ray.h"
 #include "sphere.h"
 
-Sphere *sphere_create(Vec3 c, float r) {
+Sphere *sphere_create(Vec3 c, double r) {
     Sphere *sp = malloc(sizeof(Sphere));
     sp->center = c;
     sp->radius = r;
@@ -13,13 +13,13 @@ Sphere *sphere_create(Vec3 c, float r) {
     return sp;
 }
 
-bool ray_hit_sphere_cb(Ray r, float t_min, float t_max,
+bool ray_hit_sphere_cb(Ray r, double t_min, double t_max,
                        hitrec *rec, Sphere *sp) {
     Vec3  oc;
-    float a, b, c;
-    float discriminant;
+    double a, b, c;
+    double discriminant;
 
-    float temp;
+    double temp;
 
     oc = vec3_sub(r.origin, sp->center);
     a  = vec3_dot(r.direction, r.direction);
@@ -28,7 +28,7 @@ bool ray_hit_sphere_cb(Ray r, float t_min, float t_max,
     discriminant = b*b - a*c;
 
     if (discriminant > 0) {
-        temp = (-b - sqrtf(b*b - a*c))/a;
+        temp = (-b - sqrt(b*b - a*c))/a;
         if (temp < t_max && temp > t_min) {
             rec->t = temp;
             rec->p = ray_point_at_param(r, rec->t);
@@ -37,7 +37,7 @@ bool ray_hit_sphere_cb(Ray r, float t_min, float t_max,
             return true;
         }
 
-        temp = (-b + sqrtf(b*b - a*c))/a;
+        temp = (-b + sqrt(b*b - a*c))/a;
 
         if (temp < t_max && temp > t_min) {
             rec->t = temp;
